+++ R/mock_edish.R
@@ -2,7 +2,7 @@
 #'
 #' `mock_edish_app()` launches a mock app containing the dv.edish shiny module.
 #'
-#' @keywords internal
+#' @export
 mock_edish_app <- function() {
   dm <- pharmaverseadam::adsl |> dplyr::mutate(dplyr::across(dplyr::where(is.character), as.factor))
   lb <- pharmaverseadam::adlb |> dplyr::mutate(dplyr::across(dplyr::where(is.character), as.factor))
@@ -13,9 +13,10 @@ mock_edish_app <- function() {
       arm_default_vals = "Xanomeline High Dose",
       at_choices = c("Alanine Aminotransferase", "Aspartate Aminotransferase"),
       at_default_val = "Alanine Aminotransferase",
-      tbili_choices = "Bilirubin",
+      tbili_choices = c("Bilirubin", "Eosinophils/Leukocytes"),
       tbili_default_val = "Bilirubin",
       default_by_visit = FALSE,
+      default_show_table = FALSE,
       window_days = NULL
     ))
   }
@@ -41,10 +42,13 @@ mock_edish_app <- function() {
 #'
 #' `mock_edish_mm()` launches a mock app containing the dv.edish shiny module by means of the `dv.manager`.
 #'
-#' @keywords internal
+#' @export
 mock_edish_mm <- function() {
   dm <- pharmaverseadam::adsl
   lb <- pharmaverseadam::adlb
+
+  attr(dm, "meta") <- base::file.info("NEWS.md")
+  attr(lb, "meta") <- base::file.info("NEWS.md")
 
   module_list <- list(
     "eDISH Demo" = mod_edish(