+++ NEWS.md
@@ -1,3 +1,15 @@
+# dv.edish 2.1.0
+
+- The user may switch the x- and/or y-axis between normalized values and absolute values.
+- The user can specify a ULN multiple and inequality sign (<, ≤, ≥, >) to apply baseline inclusions. This is now
+  completely independent to reference lines, and can apply to both eDISH and mDISH plots.
+- The single large drop menu has been split into 3 smaller drop menus so that less of the plot is hidden, and to reduce
+  the likelihood that the menu extends beyond the bottom of the page.
+- Implement table of subject counts and percentages in areas delimited by the specified reference lines.
+- Ensure reference line values appear on the axes.
+- Update scrollbar settings for better handling with new Shiny bootstrap.
+- Enable zoom option.
+
 # dv.edish 2.0.2
 
 - [NOT USER-FACING] Update early error feedback snippets, communication with papo test snippet and test snapshot.